18-11-2025

+++ app/Http/Controllers/Api/AuthController.php
@@ -1653,7 +1653,9 @@ public function notificationList()
                 'data' => [
                     'dp_price'          => $price ? (float)$price->dealer_price : null,
                     'adp_price'         => $price ? (float)$price->advance_dealer_price : null,
-                    'tonnage_per_piece' => $tonnage !== null ? number_format($tonnage, 4, '.', '') : null,
+                    'tonnage_per_piece' => $tonnage !== null 
+                                ? (float) number_format($tonnage, 4, '.', '') 
+                                : null,
                 ]
             ]);
